requirements.txt
alembic==1.11.1
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.74
botocore==1.31.74
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.1
cli-helpers==2.3.0
click==8.1.3
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.4
dynaconf==3.1.12
Flask==2.3.2
Flask-Bcrypt==1.0.1
flask-htmx==0.3.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gevent==23.9.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
litecli==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.2
pip-check-reqs==2.5.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.6.2
urllib3==1.26.18
wcwidth==0.2.9
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
zope.event==5.0
zope.interface==6.1
fastapi==0.103.2
uvicorn==0.23.2
azure-storage-blob==12.16.0
python-multipart==0.0.9
pytest==8.0.0
pytest-asyncio==0.23.5
